fix(spawner): move makeCollect into a separate file
This is necessary so that spawned contracts can use makeCollect
without importing all of spawner (which fails the transforms).
