+++ packages/spawner/src/contractHost.js
@@ -15,7 +15,7 @@ import { makeMint } from '@agoric/ertp';
 import makePromise from '@agoric/make-promise';
 import { makeMeter } from '@agoric/transform-metering/src/meter';
 
-import { allSettled } from './allSettled';
+export { makeCollect } from './makeCollect';
 
 /**
  * Make a reusable host that can reliably install and execute contracts.
@@ -257,32 +257,4 @@ function makeContractHost(E, evaluate, additionalEndowments = {}) {
 }
 harden(makeContractHost);
 
-function makeCollect(E, log) {
-  function collect(seatP, winPurseP, refundPurseP, name = 'collecting') {
-    const results = harden([
-      E(seatP)
-        .getWinnings()
-        .then(winnings => E(winPurseP).depositAll(winnings)),
-      // TODO Bug if we replace the comma above with the uncommented
-      // out ".then(_ => undefined)," below, somehow we end up trying
-      // to marshal an array with holes, rather than an array with
-      // undefined elements. This remains true whether we use
-      // Promise.all or allSettled
-      /* .then(_ => undefined), */
-      E(seatP)
-        .getRefund()
-        .then(refund => refund && E(refundPurseP).depositAll(refund)),
-    ]);
-    const doneP = allSettled(results);
-    Promise.resolve(doneP).then(([wins, refs]) => {
-      log(`${name} wins: `, wins, ` refs: `, refs);
-    });
-    // Use Promise.all here rather than allSettled in order to
-    // propagate rejection.
-    return Promise.all(results);
-  }
-  return harden(collect);
-}
-harden(makeCollect);
-
-export { makeContractHost, makeCollect };
+export { makeContractHost };
