+++ packages/spawner/src/makeCollect.js
@@ -1,0 +1,32 @@
+import harden from '@agoric/harden';
+import { allSettled } from './allSettled';
+
+function makeCollect(E, log) {
+  function collect(seatP, winPurseP, refundPurseP, name = 'collecting') {
+    const results = harden([
+      E(seatP)
+        .getWinnings()
+        .then(winnings => E(winPurseP).depositAll(winnings)),
+      // TODO Bug if we replace the comma above with the uncommented
+      // out ".then(_ => undefined)," below, somehow we end up trying
+      // to marshal an array with holes, rather than an array with
+      // undefined elements. This remains true whether we use
+      // Promise.all or allSettled
+      /* .then(_ => undefined), */
+      E(seatP)
+        .getRefund()
+        .then(refund => refund && E(refundPurseP).depositAll(refund)),
+    ]);
+    const doneP = allSettled(results);
+    Promise.resolve(doneP).then(([wins, refs]) => {
+      log(`${name} wins: `, wins, ` refs: `, refs);
+    });
+    // Use Promise.all here rather than allSettled in order to
+    // propagate rejection.
+    return Promise.all(results);
+  }
+  return harden(collect);
+}
+harden(makeCollect);
+
+export { makeCollect };
