+++ packages/spawner/test/swingsetTests/contractHost/vat-alice.js
@@ -3,7 +3,7 @@
 import harden from '@agoric/harden';
 import { allComparable } from '@agoric/same-structure';
 
-import { makeCollect } from '../../../src/contractHost';
+import { makeCollect } from '../../../src/makeCollect';
 
 function makeAliceMaker(E, host, log) {
   const collect = makeCollect(E, log);
